requirements.txt
# MCP Server Core
mcp>=1.7.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Web server (for SSE transport)
uvicorn>=0.27.0
fastapi>=0.109.0

# Utilities
python-dotenv>=1.0.0
